Rename `user-profile-follower-badge`
